fix(api): sync permission calc logic with sdk
i fixed some dubious permissions calculation logic in sdk, but forgot to
port it back into backend.

also:

- reduce max roles per room 1024 -> 256 (just in case)
- make role position a u16 (there won't ever be more than 64k roles lol)
- clean up some other dubious code

+++ crate-common/src/util/member_list.rs
@@ -9,7 +9,7 @@ use crate::v2::types::{RoleId, UserId, sync::subscribe::MemberListGroupId};
 #[derive(Copy, PartialEq, Eq, PartialOrd, Ord)]
 pub enum MemberGroupKey {
     Connected,
-    Hoisted { role_position: u64, role_id: RoleId },
+    Hoisted { role_position: u16, role_id: RoleId },
     Online,
     Offline,
 }
